online_cinema/requirements.txt
pytest==6.2.5
django==3.2
python-dotenv==0.19.2
django-split-settings==1.1.0
django-debug-toolbar==3.2.4
gunicorn==20.1.0
psycopg2-binary==2.9.3
pydantic==1.9.0
backoff==1.11.1
elasticsearch==7.14.1